Added hack speed input to GUI
Added an input to change the hack speed.
Cleaned up gui function - instead of having seperate variables for each element all of the gui's html is now stored into one variable.

Note: Hack value will only take come into effect when app.restart() is called. This is because the wordLoop variable is assigned the value of wordFreq in the automate function once. Can't seem to figure out how to dynamically change it.